+++ tests/flytekit/unit/remote/test_sandbox.py
@@ -1,10 +1,7 @@
 import uuid
-from datetime import timedelta
-from functools import partial
 
 import pytest
 
-from flytekit import map_task, task, workflow
 from flytekit.configuration import Config, ImageConfig, SerializationSettings
 from flytekit.loggers import logger
 from flytekit.remote import FlyteRemote
@@ -102,25 +99,3 @@ def test_run_remote_merge_sort():
     )
 
     assert exec.outputs["o0"] == [6, 9, 21, 41, 42, 71, 76, 89, 90, 94]
-
-
-@pytest.mark.sandbox_test
-def test_array_node_map():
-    # These would need to be registered.
-    @task
-    def fn(x: int, y: int) -> int:
-        return x + y
-
-    @workflow
-    def workflow_with_maptask(data: list[int], y: int) -> list[int]:
-        partial_fn = partial(fn, y=y)
-        return map_task(partial_fn)(x=data)
-
-    remote = FlyteRemote(
-        config=Config.for_sandbox(),
-        default_project="flytesnacks",
-        default_domain="development",
-    )
-
-    ex = remote.fetch_execution(name="f0c13bbb0e5084afd9b3")
-    remote.wait(ex, poll_interval=timedelta(seconds=1))
